refactor(multi-worker-demo): simplify fragment base handling and styles
