+++ multi-worker-demo/footer/src/root.tsx
@@ -4,5 +4,5 @@ import FooterCSS from "./Footer.css?inline";
 export const Footer = component$(() => {
   useStylesScoped$(FooterCSS);
 
-  return <footer class="container">Copyright Cloudflare 2022</footer>;
+  return <div class="container">Copyright Cloudflare 2022</div>;
 });
