+++ multi-worker-demo/gallery/src/Gallery.css
@@ -2,37 +2,10 @@
 	margin-top: 45px;
 }
 
-.cloud-image {
-	margin-right: 10px;
-	border-radius: 25px;
-	width: 300px;
-	height: 450px;
-}
-
-.cloud-image:last-child {
-	margin-right: 0;
-}
-
 .cloud-grid {
 	display: grid;
 	grid-template-columns: 300px 300px 300px;
 	grid-gap: 10px;
-}
-
-.image-container {
-	position: relative;
-}
-
-.tags {
-	position: absolute;
-	bottom: 15px;
-	left: 15px;
-	font-weight: 700;
-	font-size: 16px;
-	background-color: #3d3d3d;
-	color: #ededed;
-	padding: 5px;
-	border-radius: 5px;
 }
 
 .spinner-container {
