+++ multi-worker-demo/gallery/src/entry.ssr.tsx
@@ -4,14 +4,6 @@ import { renderResponse } from "helpers";
 
 export default {
   fetch(request: Request, env: Record<string, unknown>): Promise<Response> {
-    const url = new URL(request.url);
-    return renderResponse(
-      request,
-      env,
-      <Gallery base={url.searchParams.get("base") ?? ""} />,
-      // <Body />,
-      manifest,
-      "div"
-    );
+    return renderResponse(request, env, <Gallery />, manifest, "div");
   },
 };
