+++ multi-worker-demo/header/src/entry.ssr.tsx
@@ -4,14 +4,6 @@ import { renderResponse } from "helpers";
 
 export default {
   fetch(request: Request, env: Record<string, unknown>): Promise<Response> {
-    const url = new URL(request.url);
-
-    return renderResponse(
-      request,
-      env,
-      <Header base={url.searchParams.get("base") ?? ""} />,
-      manifest,
-      "header"
-    );
+    return renderResponse(request, env, <Header />, manifest, "header");
   },
 };
